add run_Selector
This is the main executable for the hlfv selector. For now it's mainly
a skeleton.  Since I was tired of parsing the options over and over, I
switched to 'getopt_long'. I still don't like its syntax, but it's
good enough for now.

Note to self: what you should implement is a thin wrapper around
'getopt_long' providing a more friendly interface, similar to the one
implemented in the fastjet parser
http://fastjet.fr/repo/doxygen-3.0.6/CmdLine_8hh_source.html
(see for example  'fastjet__timing_8.cc')