CutFlowCounter::pass() add double rather than WeightComponents
No need to depend on WeightComponents since internally I'm just
incrementing a double. For now I really don't see the need to pass in
WeightComponents: most of the times we want to do the two cutflows
(raw+weighted), and check the weight components only for a few
events. There is no need to make a cutflow for each weight component.
By avoiding the WeightComponents dependence, I can also test the
weighted counters in test_CutFlowCounter.