clarify local_output_root_file (bugfix e6ee6f1)
When I added the 'copy back syst files' I messed up the destination of
the nominal file.  The nominal file was still being copied back (on
accident) when copying back the syst ones (with the wildcard).  This
does not happen with the matrix prediction code (where there are only
weight variations).